requirements.txt
numpy
mne
pandas
matplotlib
scipy
transformers
diffusers==0.11.1
torch
torchvision
tqdm
pytorch_lightning 
h5py
PyYAML
einops
wandb
accelerate
optuna
pypygifmaker